+++ naga/src/front/wgsl/error.rs
@@ -1,3 +1,4 @@
+use crate::diagnostic_filter::FilterableTriggeringRule;
 use crate::front::wgsl::parse::directive::enable_extension::{
     EnableExtension, UnimplementedEnableExtension,
 };
@@ -194,6 +195,7 @@ pub(crate) enum Error<'a> {
     UnknownConservativeDepth(Span),
     UnknownEnableExtension(Span, &'a str),
     UnknownLanguageExtension(Span, &'a str),
+    UnknownDiagnosticRuleName(Span),
     SizeAttributeTooLow(Span, u32),
     AlignAttributeTooLow(Span, Alignment),
     NonPowerOfTwoAlignAttribute(Span),
@@ -293,6 +295,13 @@ pub(crate) enum Error<'a> {
     },
     LanguageExtensionNotYetImplemented {
         kind: UnimplementedLanguageExtension,
+        span: Span,
+    },
+    DiagnosticInvalidSeverity {
+        severity_control_name_span: Span,
+    },
+    DiagnosticNotYetImplemented {
+        triggering_rule: FilterableTriggeringRule,
         span: Span,
     },
 }
@@ -559,6 +568,15 @@ impl<'a> Error<'a> {
                 notes: vec![concat!(
                     "See available extensions at ",
                     "<https://www.w3.org/TR/WGSL/#language-extensions-sec>."
+                )
+                .into()],
+            },
+            Error::UnknownDiagnosticRuleName(span) => ParseError {
+                message: format!("unknown `diagnostic(…)` rule name `{}`", &source[span]),
+                labels: vec![(span, "not a valid diagnostic rule name".into())],
+                notes: vec![concat!(
+                    "See available trigger rules at ",
+                    "<https://www.w3.org/TR/WGSL/#filterable-triggering-rules>."
                 )
                 .into()],
             },
@@ -1006,6 +1024,38 @@ impl<'a> Error<'a> {
                         "so they can prioritize it!"
                     ),
                     kind.tracking_issue_num()
+                )],
+            },
+            Error::DiagnosticInvalidSeverity {
+                severity_control_name_span,
+            } => ParseError {
+                message: "invalid `diagnostic(…)` severity".into(),
+                labels: vec![(
+                    severity_control_name_span,
+                    "not a valid severity level".into(),
+                )],
+                notes: vec![concat!(
+                    "See available severities at ",
+                    "<https://www.w3.org/TR/WGSL/#diagnostic-severity>."
+                )
+                .into()],
+            },
+            Error::DiagnosticNotYetImplemented {
+                triggering_rule,
+                span,
+            } => ParseError {
+                message: format!(
+                    "the `{}` diagnostic filter is not yet supported",
+                    triggering_rule.to_ident()
+                ),
+                labels: vec![(span, "".into())],
+                notes: vec![format!(
+                    concat!(
+                        "Let Naga maintainers know that you ran into this at ",
+                        "<https://github.com/gfx-rs/wgpu/issues/{}>, ",
+                        "so they can prioritize it!"
+                    ),
+                    triggering_rule.tracking_issue_num()
                 )],
             },
         }